Add event "Actor: Face Player"
Adds a new event so that an actor can quickly face the player.

Although this can be done in GB Studio without this event, it is slow, inefficient, and uses up variables; this is a much more optimised approach.

The direction can be inverted so the actor faces away from the player, as well as locked to horizontal or vertical facing only as well.